Add Legacy Windows Support using WDK 7.1.0
Closes Issue: https://github.com/NSG650/NoMoreBugCheck/issues/4
