+++ NoMoreBugCheck/MAKEFILE
@@ -1,0 +1,2 @@
+!INCLUDE $(NTMAKEENV)\makefile.def
+
