+++ NoMoreBugCheck/sources
@@ -1,0 +1,4 @@
+TARGETNAME=nomorebugcheck
+TARGETTYPE=DRIVER
+SOURCES=driver.cpp
+MSC_WARNING_LEVEL=/W0
